Updated the python script name just to remove the white space making directory work easier
Explicitly added variable names to remove errors seen by FrancoisGLEYZON when compiling the script
This should create the base set of Transient variable names when transient variables are used

I also fixed the convective boundary conditions in the example transient file. I had them in fahrenheit instead of Kelvin
